test: corrected acceptance tests to handle GeneratingTriples event status
