+++ event-log/src/test/scala/io/renku/eventlog/InMemoryEventLogDbSpec.scala
@@ -30,7 +30,11 @@ import doobie.implicits._
 import doobie.util.fragment.Fragment
 import org.scalatest.TestSuite
 
-trait InMemoryEventLogDbSpec extends DbSpec with InMemoryEventLogDb with EventLogDataProvisioning {
+trait InMemoryEventLogDbSpec
+    extends DbSpec
+    with InMemoryEventLogDb
+    with EventLogDataProvisioning
+    with EventLogDataFetching {
   self: TestSuite =>
 
   protected def initDb(): Unit = {
@@ -65,76 +69,4 @@ trait InMemoryEventLogDbSpec extends DbSpec with InMemoryEventLogDb with EventLo
       Fragment.const(s"TRUNCATE TABLE $tableName").update.run.map(_ => ())
     }
   }
-
-  protected def storeEvent(compoundEventId: CompoundEventId,
-                           eventStatus:     EventStatus,
-                           executionDate:   ExecutionDate,
-                           eventDate:       EventDate,
-                           eventBody:       EventBody,
-                           createdDate:     CreatedDate = CreatedDate(Instant.now),
-                           batchDate:       BatchDate = BatchDate(Instant.now),
-                           projectPath:     Path = projectPaths.generateOne,
-                           maybeMessage:    Option[EventMessage] = None
-  ): Unit = {
-    insertEvent(compoundEventId, eventStatus, executionDate, eventDate, eventBody, createdDate, batchDate, maybeMessage)
-    upsertProject(compoundEventId, projectPath, eventDate)
-  }
-
-  protected def insertEvent(compoundEventId: CompoundEventId,
-                            eventStatus:     EventStatus,
-                            executionDate:   ExecutionDate,
-                            eventDate:       EventDate,
-                            eventBody:       EventBody,
-                            createdDate:     CreatedDate,
-                            batchDate:       BatchDate,
-                            maybeMessage:    Option[EventMessage]
-  ): Unit = execute {
-    maybeMessage match {
-      case None =>
-        sql"""|INSERT INTO
-              |event (event_id, project_id, status, created_date, execution_date, event_date, batch_date, event_body)
-              |VALUES (${compoundEventId.id}, ${compoundEventId.projectId}, $eventStatus, $createdDate, $executionDate, $eventDate, $batchDate, $eventBody)
-      """.stripMargin.update.run.map(_ => ())
-      case Some(message) =>
-        sql"""|INSERT INTO
-              |event (event_id, project_id, status, created_date, execution_date, event_date, batch_date, event_body, message)
-              |VALUES (${compoundEventId.id}, ${compoundEventId.projectId}, $eventStatus, $createdDate, $executionDate, $eventDate, $batchDate, $eventBody, $message)
-      """.stripMargin.update.run.map(_ => ())
-    }
-  }
-
-  // format: off
-  protected def findEvents(status:  EventStatus,
-                           orderBy: Fragment = fr"created_date asc"): List[(CompoundEventId, ExecutionDate, BatchDate)] =
-    execute {
-      (fr"""SELECT event_id, project_id, execution_date, batch_date
-            FROM event
-            WHERE status = $status
-            ORDER BY """ ++ orderBy)
-        .query[(CompoundEventId, ExecutionDate, BatchDate)]
-        .to[List]
-    }
-  // format: on
-
-  protected def findProjects(): List[(projects.Id, projects.Path, EventDate)] = execute {
-    sql"""SELECT * FROM project"""
-      .query[(projects.Id, projects.Path, EventDate)]
-      .to[List]
-  }
-
-  protected implicit class FoundEventsOps(events: List[(CompoundEventId, ExecutionDate, BatchDate)]) {
-    lazy val noBatchDate: List[(CompoundEventId, ExecutionDate)] = events.map { case (id, executionDate, _) =>
-      id -> executionDate
-    }
-    lazy val eventIdsOnly: List[CompoundEventId] = events.map { case (id, _, _) => id }
-  }
-
-  protected def findEventMessage(eventId: CompoundEventId): Option[EventMessage] =
-    execute {
-      sql"""SELECT message
-            FROM event 
-            WHERE event_id = ${eventId.id} AND project_id = ${eventId.projectId}"""
-        .query[Option[EventMessage]]
-        .unique
-    }
 }
